+++ benchopt/cli/main.py
@@ -2,10 +2,10 @@ import click
 from pathlib import Path
 
 from benchopt.benchmark import Benchmark
-from benchopt.cli.completion import get_solvers
-from benchopt.cli.completion import get_datasets
-from benchopt.cli.completion import get_benchmark
-from benchopt.cli.completion import get_conda_envs
+from benchopt.cli.completion import complete_solvers
+from benchopt.cli.completion import complete_datasets
+from benchopt.cli.completion import complete_benchmarks
+from benchopt.cli.completion import complete_conda_envs
 from benchopt.utils.conda_env_cmd import list_conda_envs
 from benchopt.utils.conda_env_cmd import create_conda_env
 from benchopt.utils.shell_cmd import _run_shell_in_conda_env
@@ -26,7 +26,7 @@ main = click.Group(
     "conda environment, see the command `benchopt install`."
 )
 @click.argument('benchmark', type=click.Path(exists=True),
-                autocompletion=get_benchmark)
+                shell_complete=complete_benchmarks)
 @click.option('--objective-filter', '-p', 'objective_filters',
               metavar='<objective_filter>', multiple=True, type=str,
               help="Filter the objective based on its parametrized name. This "
@@ -36,26 +36,26 @@ main = click.Group(
               help="Include <solver_name> in the benchmark. By default, all "
               "solvers are included. When `-s` is used, only listed solvers"
               " are included. To include multiple solvers, "
-              "use multiple `-s` options.", autocompletion=get_solvers)
+              "use multiple `-s` options.", shell_complete=complete_solvers)
 @click.option('--force-solver', '-f', 'forced_solvers',
               metavar="<solver_name>", multiple=True, type=str,
               help="Force the re-run for <solver_name>. This "
-              "avoids caching effect when adding a solver."
+              "avoids caching effect when adding a solver. "
               "To select multiple solvers, use multiple `-f` options.",
-              autocompletion=get_solvers)
+              shell_complete=complete_solvers)
 @click.option('--dataset', '-d', 'dataset_names',
               metavar="<dataset_name>", multiple=True, type=str,
               help="Run the benchmark on <dataset_name>. By default, all "
               "datasets are included. When `-d` is used, only listed datasets"
               " are included. Note that <dataset_name> can be a regexp. "
               "To include multiple datasets, use multiple `-d` options.",
-              autocompletion=get_datasets)
+              shell_complete=complete_datasets)
 @click.option('--max-runs', '-n',
               metavar="<int>", default=100, show_default=True, type=int,
               help='Maximal number of runs for each solver. This corresponds '
               'to the number of points in the time/accuracy curve.')
 @click.option('--n-repetitions', '-r',
-              metavar='<int>', default=5, type=int,
+              metavar='<int>', default=5, show_default=True, type=int,
               help='Number of repetitions that are averaged to estimate the '
               'runtime.')
 @click.option('--timeout',
@@ -86,7 +86,7 @@ main = click.Group(
               "benchopt_<BENCHMARK>.")
 @click.option('--env-name', 'env_name',
               metavar="<env_name>", type=str, default='False',
-              autocompletion=get_conda_envs,
+              shell_complete=complete_conda_envs,
               help="Run the benchmark in the conda environment "
               "named <env_name>. To install the required solvers and "
               "datasets, see the command `benchopt install`.")
@@ -105,6 +105,7 @@ def run(benchmark, solver_names, forced_solvers, dataset_names,
     benchmark = Benchmark(benchmark)
     benchmark.validate_dataset_patterns(dataset_names)
     benchmark.validate_solver_patterns(solver_names+forced_solvers)
+    benchmark.validate_objective_filters(objective_filters)
 
     # If env_name is False, the flag `--local` has been used (default) so
     # run in the current environement.
@@ -156,9 +157,9 @@ def run(benchmark, solver_names, forced_solvers, dataset_names,
 
     # run the command in the conda env
     solvers_option = ' '.join(['-s ' + s for s in solver_names])
-    forced_solvers_option = ' '.join(['-f ' + s for s in forced_solvers])
-    datasets_option = ' '.join(['-d ' + d for d in dataset_names])
-    objective_option = ' '.join(['-p ' + p for p in objective_filters])
+    forced_solvers_option = ' '.join([f"-f '{s}'" for s in forced_solvers])
+    datasets_option = ' '.join([f"-d '{d}'" for d in dataset_names])
+    objective_option = ' '.join([f"-p '{p}'" for p in objective_filters])
     cmd = (
         rf"benchopt run --local {benchmark.benchmark_dir} "
         rf"--n-repetitions {n_repetitions} "
@@ -179,7 +180,7 @@ def run(benchmark, solver_names, forced_solvers, dataset_names,
     help="Install the requirements (solvers/datasets) for a benchmark."
 )
 @click.argument('benchmark', type=click.Path(exists=True),
-                autocompletion=get_benchmark)
+                shell_complete=complete_benchmarks)
 @click.option('--force', '-f',
               is_flag=True,
               help="If this flag is set, the reinstallation of "
@@ -187,17 +188,23 @@ def run(benchmark, solver_names, forced_solvers, dataset_names,
 @click.option('--solver', '-s', 'solver_names',
               metavar="<solver_name>", multiple=True, type=str,
               help="Include <solver_name> in the installation. "
-              "By default, all solvers are included. "
+              "By default, all solvers are included except "
+              "when -d flag is used. If -d flag is used, then "
+              "no solver is included by default. "
               "When `-s` is used, only listed estimators are included. "
-              "To include multiple solvers, use multiple `-s` options.",
-              autocompletion=get_solvers)
+              "To include multiple solvers, use multiple `-s` options."
+              "To include all solvers, use -s 'all' option.",
+              shell_complete=complete_solvers)
 @click.option('--dataset', '-d', 'dataset_names',
               metavar="<dataset_name>", multiple=True, type=str,
               help="Install the dataset <dataset_name>. By default, all "
-              "datasets are included. When `-d` is used, only listed datasets "
+              "datasets are included, except when -s flag is used. "
+              "If -s flag is used, then no dataset is included. "
+              "When `-d` is used, only listed datasets "
               "are included. Note that <dataset_name> can be a regexp. "
-              "To include multiple datasets, use multiple `-d` options.",
-              autocompletion=get_datasets)
+              "To include multiple datasets, use multiple `-d` options."
+              "To include all datasets, use -d 'all' option.",
+              shell_complete=complete_datasets)
 @click.option('--env', '-e', 'env_name',
               flag_value='True', type=str, default='False',
               help="Install all requirements in a dedicated "
@@ -206,7 +213,7 @@ def run(benchmark, solver_names, forced_solvers, dataset_names,
               "solver dependencies and datasets are installed in it.")
 @click.option('--env-name', 'env_name',
               metavar="<env_name>", type=str, default='False',
-              autocompletion=get_conda_envs,
+              shell_complete=complete_conda_envs,
               help="Install the benchmark requirements in the "
               "conda environment named <env_name>. If it does not exist, "
               "it will be created by this command.")
@@ -214,12 +221,15 @@ def run(benchmark, solver_names, forced_solvers, dataset_names,
               help="If this flag is set, start with a fresh conda "
               "environment. It can only be used combined with options "
               "`-e/--env` or `--env-name`.")
+@click.option('--quiet', '-q', 'quiet', is_flag=True, default=False,
+              show_default=True,
+              help="If this flag is set, conda's output is silenced.")
 @click.option('--yes', '-y', 'confirm', is_flag=True,
               help="If this flag is set, no confirmation will be asked "
               "to the user to install requirements in the current environment."
               " Useless with options `-e/--env` or `--env-name`.")
 def install(benchmark, solver_names, dataset_names, force=False,
-            recreate=False, env_name='False', confirm=False):
+            recreate=False, env_name='False', confirm=False, quiet=False):
 
     # Check that the dataset/solver patterns match actual dataset
     benchmark = Benchmark(benchmark)
@@ -267,13 +277,13 @@ def install(benchmark, solver_names, dataset_names, force=False,
                 )
 
         # create environment if necessary
-        create_conda_env(env_name, recreate=recreate)
+        create_conda_env(env_name, recreate=recreate, quiet=quiet)
 
     # install requirements
     print("# Install", flush=True)
     benchmark.install_all_requirements(
         include_solvers=solver_names, include_datasets=dataset_names,
-        env_name=env_name, force=force
+        env_name=env_name, force=force, quiet=quiet,
     )
 
 
@@ -282,9 +292,9 @@ def install(benchmark, solver_names, dataset_names, force=False,
     context_settings=dict(ignore_unknown_options=True)
 )
 @click.argument('benchmark', type=click.Path(exists=True),
-                autocompletion=get_benchmark)
+                shell_complete=complete_benchmarks)
 @click.option('--env-name', type=str, default=None, metavar='NAME',
-              autocompletion=get_conda_envs,
+              shell_complete=complete_conda_envs,
               help='Environment to run the test in. If it is not provided '
               'a temporary one is created for the test.')
 @click.argument('pytest_args', nargs=-1, type=click.UNPROCESSED)