+++ benchopt/utils/colorify.py
@@ -1,8 +1,12 @@
 "Helper function for colored terminal outputs"
+import shutil
 
 
-LINE_LENGTH = 100
+MIN_LINE_LENGTH = 20
 BLACK, RED, GREEN, YELLOW, BLUE, MAGENTA, CYAN, WHITE = range(30, 38)
+
+CROSS = u'\u2717'
+TICK = u'\u2713'
 
 
 def colorify(message, color=BLUE):
@@ -18,4 +22,18 @@ def colorify(message, color=BLUE):
     color_message : str
         The colored message to be displayed in terminal.
     """
-    return ("\033[1;%dm" % color) + message + "\033[0m"
+    return f"\033[1;{color}m" + message + "\033[0m"
+
+
+def print_normalize(msg, endline=True):
+    """Format the output to have the length of the terminal."""
+    line_length = max(
+        MIN_LINE_LENGTH, shutil.get_terminal_size((100, 24)).columns
+    )
+    n_colors = msg.count('\033') // 2
+    msg = msg.ljust(line_length + n_colors * 11)
+
+    if endline:
+        print(msg)
+    else:
+        print(msg + '\r', end='', flush=True)