+++ doc/whats_new.rst
@@ -13,6 +13,9 @@ Version 1.2 - in progress
 Changelog
 ~~~~~~~~~
 
+- New ``benchopt info`` command to display information about solvers and datasets
+  of a benchmark, by `Ghislain Durif`_ (:gh:`140`).
+
 - New ``--profile`` option to the ``run`` command in order to profile
   with the line-profiler package all functions decorated with
   :func:`benchopt.utils.profile`, by `Alexandre Gramfort`_ (:gh:`186`).
@@ -22,15 +25,17 @@ Changelog
   the previous one, by `Thomas Moreau`_ (:gh:`176`)
 
 - Now all values returned by ``Objective.compute`` are included in reports,
-  by `Thomas Moreau`_ and `Alexandre Gramfort`_ (gh:`200`).
+  by `Thomas Moreau`_ and `Alexandre Gramfort`_ (:gh:`200`).
 
 API
 ~~~
 
 - When returning a dict, ``Objective.compute`` should at least include
   ``value`` key instead of ``objective_value``, by `Thomas Moreau`_ and
-  `Alexandre Gramfort`_ (gh:`200`).
+  `Alexandre Gramfort`_ (:gh:`200`).
 
+- 'stop_strategy' attribute is replaced by 'stopping_strategy' to harmonize
+  with 'stopping_criterion', by `Benoît Malézieux`_ (:gh:`274`).
 
 .. _changes_1_1:
 
@@ -64,7 +69,7 @@ API
 ~~~
 
 - ``Objective.compute`` can now return a dictionary with multiple outputs to
-  monitor several metrics at once, by `Thomas Moreau`_ (gh:`84`).
+  monitor several metrics at once, by `Thomas Moreau`_ (:gh:`84`).
 
 - ``Solver.skip`` can now be used to skip objectives that are incompatible
   for the Solver, by `Thomas Moreau`_ (:gh:`113`).
@@ -102,10 +107,15 @@ CLI
 - Add ``benchopt install`` command to install benchmark requirements
   (not done in ``benchopt run`` anymore) by `Ghislain Durif`_ (:gh:`135`).
 
+- Add ``benchopt info`` command to print information about a benchmark
+  (including solvers, datasets, dependencies, etc.)
+  by `Ghislain Durif`_ (:gh:`140`).
+
+
 BUG
 ~~~
 
-- Throw a warning when BenchOpt version in conda env does not match the one of
+- Throw a warning when benchopt version in conda env does not match the one of
   calling ``benchopt``, by `Thomas Moreau`_ (:gh:`83`).
 
 - Fix Lapack issue with R code, by `Tanguy Lefort`_ (:gh:`97`).
@@ -123,7 +133,7 @@ DOC
 The committer list for this release is the following:
 
   * `Alexandre Gramfort`_
-  * `Benoit Malézieux`_
+  * `Benoît Malézieux`_
   * `Ghislain Durif`_
   * `Joseph Salmon`_
   * `Mathurin Massias`_